+++ common/src/main/java/me/pandamods/fallingtrees/config/FallingTreesConfig.java
@@ -1,17 +1,17 @@
 package me.pandamods.fallingtrees.config;
 
 import me.pandamods.fallingtrees.FallingTrees;
-import me.pandamods.pandalib.config.api.ConfigRegistry;
-import me.pandamods.pandalib.config.api.holders.ClientConfigHolder;
-import me.pandamods.pandalib.config.api.holders.CommonConfigHolder;
+import me.pandamods.pandalib.api.config.PandaLibConfig;
+import me.pandamods.pandalib.api.config.holders.ClientConfigHolder;
+import me.pandamods.pandalib.api.config.holders.CommonConfigHolder;
 import net.minecraft.world.entity.player.Player;
 
 public class FallingTreesConfig {
-	public final ClientConfigHolder<ClientConfig> clientConfigHolder = ConfigRegistry.registerClient(ClientConfig.class);
-	public final CommonConfigHolder<CommonConfig> commonConfigHolder = ConfigRegistry.registerCommon(CommonConfig.class);
+	public final ClientConfigHolder<ClientConfig> clientConfigHolder = PandaLibConfig.registerClient(ClientConfig.class);
+	public final CommonConfigHolder<CommonConfig> commonConfigHolder = PandaLibConfig.registerCommon(CommonConfig.class);
 
 	public static ClientConfig getClientConfig(Player player) {
-		return FallingTrees.CONFIG.clientConfigHolder.getClient(player);
+		return player.pandaLib$getConfig(FallingTrees.CONFIG.clientConfigHolder);
 	}
 
 	public static ClientConfig getClientConfig() {