+++ common/src/main/java/me/pandamods/fallingtrees/config/common/tree/StandardTreeConfig.java
@@ -1,6 +1,7 @@
 package me.pandamods.fallingtrees.config.common.tree;
 
 import net.minecraft.tags.BlockTags;
+import net.minecraft.tags.ItemTags;
 import net.minecraft.world.level.block.Blocks;
 
 import java.util.ArrayList;
@@ -10,18 +11,27 @@ public class StandardTreeConfig extends TreeConfig {
 	public Algorithm algorithm = new Algorithm();
 
 	public Filter logFilter = new Filter(
-			List.of(BlockTags.LOGS),
+			List.of(BlockTags.LOGS.location().toString()),
 			new ArrayList<>(),
 			new ArrayList<>()
 	);
 	public Filter leavesFilter = new Filter(
-			List.of(BlockTags.LEAVES),
+			List.of(BlockTags.LEAVES.location().toString()),
 			new ArrayList<>(),
 			new ArrayList<>()
 	);
 	public Filter extraBlockFilter = new Filter(
 			new ArrayList<>(),
-			List.of(Blocks.VINE, Blocks.BEE_NEST, Blocks.COCOA),
+			List.of(
+					Blocks.VINE.arch$registryName().toString(),
+					Blocks.BEE_NEST.arch$registryName().toString(),
+					Blocks.COCOA.arch$registryName().toString()
+			),
+			new ArrayList<>()
+	);
+	public Filter allowedToolFilter = new Filter(
+			List.of(ItemTags.AXES.location().toString()),
+			new ArrayList<>(),
 			new ArrayList<>()
 	);
 
@@ -29,5 +39,6 @@ public class StandardTreeConfig extends TreeConfig {
 		public int maxLeavesRadius = 10;
 		public int maxLogAmount = 256;
 		public boolean shouldFallOnMaxLogAmount = false;
+		public boolean shouldIgnorePersistentLeaves = true;
 	}
 }